Move annotations to applicationmodel folder - inline with aspire repo
