+++ src/CommunityToolkit.Aspire.Hosting.Dapr/ApplicationModel/DaprComponentReferenceAnnotation.cs
