+++ src/CommunityToolkit.Aspire.Hosting.Dapr/ApplicationModel/DaprSidecarAnnotation.cs
@@ -1,9 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Aspire.Hosting.ApplicationModel;
-
-namespace Aspire.Hosting;
+namespace Aspire.Hosting.ApplicationModel;
 
 /// <summary>
 /// Indicates that a Dapr sidecar should be started for the associated resource.
