+++ src/CommunityToolkit.Aspire.Hosting.Dapr/ApplicationModel/DaprSidecarOptionsAnnotation.cs
@@ -1,9 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Aspire.Hosting.ApplicationModel;
-
-namespace Aspire.Hosting;
+namespace Aspire.Hosting.ApplicationModel;
 
 /// <summary>
 /// Indicates the options used to configure a Dapr sidecar.
